+++ middleware/authMiddleware.js
@@ -1,0 +1,13 @@
+module.exports.isAuth = (req, res, next) => {
+  if (req.isAuthenticated()) {
+    return next();
+  }
+  res.redirect("/login");
+};
+
+module.exports.isNotAuth = (req, res, next) => {
+  if (req.isAuthenticated()) {
+    return res.redirect("/");
+  }
+  next();
+};
