+++ routes/index.js
@@ -6,8 +6,8 @@ const passport = require("passport");
 const flash = require("express-flash");
 const session = require("express-session");
 const initializePassport = require("../passport-config");
-const isAuth = require("./authMiddleware").isAuth;
-const isNotAuth = require("./authMiddleware").isNotAuth;
+const isAuth = require("../middleware/authMiddleware").isAuth;
+const isNotAuth = require("../middleware/authMiddleware").isNotAuth;
 
 initializePassport(passport, getUserByUsername, getUserById);
